refactor(hw/motherboard): split files
